Move export image to top bar as print
