+++ frontend/src/components/Common/ReportDialog/types.tsx
@@ -1,0 +1,4 @@
+export enum ReportType {
+  Storm,
+  Flood,
+}
