+++ frontend/src/components/MapView/index.test.tsx
@@ -12,7 +12,6 @@ jest.mock('./Layers/BoundaryLayer', () => 'mock-BoundaryLayer');
 
 jest.mock('./Legends', () => 'mock-Legends');
 jest.mock('./DateSelector', () => 'mock-DateSelector');
-jest.mock('./Download', () => 'mock-Download');
 
 jest.mock('react-router-dom', () => ({
   useHistory: () => ({
