+++ frontend/src/components/NavBar/PrintImage/index.test.tsx
